Other Function

+++ src/main/java/domain/Controllers/AssociationRepresentativeController.java
@@ -7,8 +7,10 @@ import domain.SeasonPolicies.AssignPolicy;
 import domain.SeasonPolicies.ScoreComputingPolicy;
 import domain.Users.AssociationRepresentative;
 import domain.Users.Referee;
+import domain.Users.SignedUser;
 import org.apache.commons.validator.routines.EmailValidator;
 
+import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -20,12 +22,17 @@ public class AssociationRepresentativeController {
     public boolean defineNewLeague(AssociationRepresentative associationRepresentative, String leagueName, RefereeTraining refereeTraining) throws Exception {
         if (!leaguesNameValidation(leagueName))
             throw new Exception("This league name is already exist");
+
         League newLeague = new League(leagueName, associationRepresentative, refereeTraining);
         addNewLeague(leagueName, newLeague);
+
+        //Logger
+        SystemController.logger.info("New league have been defined; Association Representative ID: " + associationRepresentative.getAssociationRepresentativeID() +
+                "; League ID: " + newLeague.getLeagueID() + "; League Name:" + newLeague.getLeagueName());
+
         return true;
     }
 
-    //TODO
     //Use Case 9.2
     public boolean addSeasonToLeague(AssociationRepresentative associationRepresentative, League league, Integer year, long startDate) throws Exception {
     int length = String.valueOf(year).length();
@@ -35,75 +42,107 @@ public class AssociationRepresentativeController {
     if(!league.checkForSeason(year)){
         throw new Exception("There is season in this year for this league");
     }
-    //TODO ADD TO THE USE CASE
+
     Season season = new Season(year, startDate);
     league.addSeason(year,season);
+
+    //Logger
+     SystemController.logger.info("New Season have been add to league; season ID: " + season.getSeasonID() +
+                "; League ID: " + league.getLeagueID() + "; Association Representative ID:" + associationRepresentative.getAssociationRepresentativeID());
+
     return true;
     }
 
     //Use Case 9.3.1
-    public boolean appointReferee (AssociationRepresentative associationRepresentative, Integer id, String fName, String lName, String email, RefereeTraining refereeTraining ) throws Exception {
+    public boolean appointReferee (AssociationRepresentative associationRepresentative, int id, String fName, String lName, String email, RefereeTraining refereeTraining ) throws Exception {
         int length = String.valueOf(id).length();
+        String userName = fName+"_"+lName;
+        String password = String.valueOf(id);
         boolean valid = EmailValidator.getInstance().isValid(email);
+
         if(length != 9)
             throw new Exception("Not Valid ID");
         if(!valid)
             throw new Exception("Not Valid Email");
-        if(IDReferees.containsKey(id))
+
+        if(findUserByID(id))
             throw new Exception("This ID Exist In The System");
 
-        String userName = String.valueOf(id);
-        String password = String.valueOf(id);
-        //TODO Send Email
+        boolean send = ExternalServices.sendInviteToTheSystem(email, userName, password, associationRepresentative.getFirstName() + " " + associationRepresentative.getLastName());
+        if(!send)
+            throw new Exception("Have been problem with send the email");
 
         String hashPassword = Utils.sha256(password);
         Referee newReferee = new Referee(userName,hashPassword,id,fName,lName,email,refereeTraining);
-        IDReferees.put(id,newReferee);
+        userNameUser.put(userName,newReferee);
+
+        //Logger
+        SystemController.logger.info("New Referee have been appoint; referee ID: " + newReferee.getRefereeID() +
+                "; Association Representative ID:" + associationRepresentative.getAssociationRepresentativeID());
+
         return true;
     }
-
 
     //Use Case 9.3.2
     //TODO need to find replacer referee
     public boolean removeReferee (AssociationRepresentative associationRepresentative, Referee referee) throws Exception {
+        SignedUser remove = userNameUser.remove(referee.getFirstName() + "_" + referee.getLastName());
+        if(remove==null){
+            throw new Exception("This Referee use name doesn't exist in the system");
+        }
         referee.deleteUser();
         archiveUsers.add(referee);
+        //Logger
+        SystemController.logger.info("Referee have been remove from the system; Referee ID: " + referee.getRefereeID() +
+                "; Association Representative ID:" + associationRepresentative.getAssociationRepresentativeID());
         return true;
     }
 
     //Use Case 9.4 A
     public Set<Referee> getAllRefereeThatCanBeForLeague(League league){
         int numTraining = league.getMinRefereeTrainingRequired().getNumVal();
-        Set<Referee> relevantReferees = IDReferees.values().stream().filter(referee -> referee.getRefereeTraining().getNumVal() <= numTraining).collect(Collectors.toSet());
-        return relevantReferees;
+        HashSet<Referee> refereeThatFitToTraining = findRefereeThatFitToTraining(numTraining);
+        return refereeThatFitToTraining;
     }
 
     //Use Case 9.4 B
-    public boolean setRefereeToLeague(Season season, Referee... referees){
+    public boolean setRefereeToLeague(AssociationRepresentative associationRepresentative, Season season, Referee... referees){
         for (Referee referee : referees) {
             season.addReferee(referee.getRefereeTraining(), referee);
             referee.addSeason(season);
+            //Logger
+            SystemController.logger.info("New Referee have been appoint to season; SeasonID: "+ season.getSeasonID()+"; Referee ID: " + referee.getRefereeID() +
+                    "; Association Representative ID:" + associationRepresentative.getAssociationRepresentativeID());
         }
         return true;
     }
 
     //UseCase 9.5
-    public boolean setScoreComputingPolicy(Season season, ScoreComputingPolicy scoreComputingPolicy) throws Exception {
+    public boolean setScoreComputingPolicy(AssociationRepresentative associationRepresentative, Season season, ScoreComputingPolicy scoreComputingPolicy) throws Exception {
         long currentTime =System.currentTimeMillis();
         if(season.getStartDate()<=currentTime)
             throw new Exception("Season already started");
 
         season.setScorePolicy(scoreComputingPolicy);
+
+        //Logger
+        SystemController.logger.info("New Score Computing have been set to season; SeasonID: "+ season.getSeasonID()+"; Score Computing Name: " + scoreComputingPolicy.getName() +
+                "; Association Representative ID:" + associationRepresentative.getAssociationRepresentativeID());
+
         return true;
 
     }
 
     //UseCase 9.6
-    public boolean setAssignPolicy(Season season, AssignPolicy assignPolicy) throws Exception {
+    public boolean setAssignPolicy(AssociationRepresentative associationRepresentative, Season season, AssignPolicy assignPolicy) throws Exception {
         long currentTime =System.currentTimeMillis();
         if(season.getStartDate()<=currentTime)
             throw new Exception("Season already started");
         season.setAssignPolicy(assignPolicy);
+
+        //Logger
+        SystemController.logger.info("New Assign Policy have been set to season; SeasonID: "+ season.getSeasonID()+"; Assign Policy Name: " + assignPolicy.getName() +
+                "; Association Representative ID:" + associationRepresentative.getAssociationRepresentativeID());
         return true;
 
     }
